+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gigadrive/sdk
 
+## 0.1.2
+
+### Patch Changes
+
+- update dependencies ([#232](https://github.com/Gigadrive/sdk/pull/232))
+
 ## 0.1.1
 
 ### Patch Changes